Add files via upload
In the Freight Transport project, the Terminal microservice is responsible for handling all terminal-related functionalities. This includes managing terminal locations, tracking terminal operations, and coordinating with other microservices to ensure smooth transportation of goods. The code related to this microservice is developed using the Spring framework and follows the principles of microservices architecture, such as modularity, scalability, and loose coupling. The Terminal microservice communicates with other microservices through RESTful API calls.